WIP Start mapping e500 itinerary model to new scheduling model
See comment on `class WriteRes` in llvm/include/llvm/Target/TargetSchedule.td
for a starting point on how to map the old (itinerary) model to the new
scheduling model. As far as I've seen ARMScheduleA9.td is the only cpu model
currently doing this mapping so this work is based off of that and the existing
info in the e500 model.